Rainfall and PET Analysis of Kars
Panoply was used to visualize the .nc files downloaded from ERA-5 websites. CDO was utilized to
extract monthly and yearly PET and P values. To identify the driest-wettest months and years of the
data, P values were compared. Mann-Kendall Trend Test is applied with XLSTAT on Excel. The Thornthwaite-Mather Water Balance Graph was created using Python